Add common itrans bindings found in itranslator (see http://www.omkarananda-ashram.org/Sanskrit/i99help/index.html?itrans_5_30.htm)
Add simplified bindings so that punctuation '^' and '~' are not required.
R and L are used for R^i and L^i.
Y is used for ~j allowing for jY in place of j~j or jJY.

Current bindings for R and Y are removed
(R has alternate bindings, .Y is used in place of Y).